requirements.txt
attrdict==2.0.1
fastapi==0.89.1
matplotlib==3.5.2
numpy==1.23.5
Pillow==9.4.0
pycocotools==2.0.6
PyYAML==6.0
torch==1.13.0.dev20220625
torchvision==0.14.0.dev20220625
tqdm==4.64.1
opencv-python
